refactor: - поправил лишнее оборачивание в массив в редьюсере для удаления и переключения эдемента списка - вынес редьюсер и его типы в src/store
style:
- удалил неиспользуемые классы и стили
