+++ src/components/Form/Form.module.css
@@ -1,5 +1,6 @@
 .formContainer {
     width: 350px;
+    margin-bottom: 15px;
 }
 
 .form {
