--- src/components/List/List.module.css
@@ -1,3 +1,0 @@
-.formWrapper {
-    margin-bottom: 15px;
-}
